feat: add district house of reps page
